pyproject.toml
[tool.poetry]
name = "core-app"
version = "0.1.0"
description = ""
authors = ["unimaginablecat <monkok111@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
psycopg2 = "^2.9.9"
django = "^5.1.2"
djangorestframework = "^3.15.2"
pytest-django = "^4.9.0"
drf-yasg = "^1.21.7"
setuptools = "^75.2.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
mypy = "^1.12.0"
ruff = "^0.6.9"


[tool.ruff]
target-version = "py312"
line-length = 88
lint.select = ["E", "F", "W", "C90"]
exclude = [
    "core_app/settings.py",
    "core_app/asgi.py",
    "core_app/wsgi.py",
    "manage.py",
    "start_script.py",
    "users_app/migrations",
]
lint.ignore = ["E501", "F401"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start_script = "start_script:main"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core_app.settings"